+++ src/components/InteractionLayout/NavigationBar/NavigationBar.jsx
@@ -1,8 +1,28 @@
-import { AppBar, TitlePortal } from "react-admin";
+import { AppBar, TitlePortal, useRefresh } from "react-admin";
 import config from "../../../config";
 import "./NavigationBar.css";
 import AuthenticationMenu from "../AuthenticationMenu/AuthenticationMenu";
 import { Component } from "react";
+import SparqlDataProvider from "./../../../dataProvider/SparqlDataProvider";
+import CleaningServicesIcon from '@mui/icons-material/CleaningServices';
+import { IconButton } from '@mui/material';
+import { Tooltip } from '@mui/material';
+
+
+function InvalidateButton() {
+  const refresh = useRefresh();
+  const handleClick = () => {
+    SparqlDataProvider.queryEngine.invalidateHttpCache();
+    setTimeout(refresh, 2000);
+  }
+  return (
+    <Tooltip title="Clean Query Cache">
+      <IconButton color="inherit" onClick={handleClick}>
+        <CleaningServicesIcon />
+      </IconButton>
+    </Tooltip>
+  )  
+}
 
 /**
  * 
@@ -18,6 +38,7 @@ function NavigationBar(props) {
         alt="Web application logo"
       ></img>
       <TitlePortal/>
+      <InvalidateButton/>
     </AppBar>
   );
 }